+++ src/lib.rs
@@ -1,38 +1,80 @@
-use std::{path::PathBuf, process::Command};
+use anyhow::{Result, anyhow};
 use serde_json::Value;
-use anyhow::{anyhow, Result};
+use std::{path::PathBuf, process::Command};
 
-struct Inventory {
-    path: PathBuf,
-    data: Value,
+const ANSIBLE_INVENTORY_COMMAND: &str = "ansible-inventory";
+
+pub struct Inventory {
+    pub path: PathBuf,
+    pub data: Value,
 }
 
-struct Host {
-    name: String,
-    data: Value,
+#[derive(Debug, Clone)]
+pub struct Host {
+    pub name: Value,
+    pub data: Value,
 }
 
-impl Inventory {
-    fn load(self) -> Result<Value> {
-        // configure the command
-        // execute the command
-        // save output to Ok(json)
-        // return
-        todo!();
-        return Err(anyhow!("-1"))
+pub trait Load
+where
+    Self: std::marker::Sized,
+{
+    fn load(source: PathBuf) -> Result<Self>;
+}
+
+impl Load for Inventory {
+    fn load(source: PathBuf) -> Result<Inventory> {
+        assert!(
+            Command::new("which")
+                .arg(ANSIBLE_INVENTORY_COMMAND)
+                .output()
+                .unwrap()
+                .status
+                .success()
+        );
+        if !source.exists() {
+            return Err(anyhow!("Invalid path to Ansible inventory: {:?}", source));
+        }
+        let command = Command::new(ANSIBLE_INVENTORY_COMMAND)
+            .args([
+                "--inventory",
+                source
+                    .to_str()
+                    .expect("Inventory source is not valid unicode"),
+                "--list",
+            ])
+            .output()?;
+        if command.status.success() {
+            let raw_data = String::from_utf8_lossy(&command.stdout);
+            let data: Value = serde_json::from_str(&raw_data).unwrap();
+            Ok(Inventory { path: source, data })
+        } else {
+            Err(anyhow!(
+                "Ansible command failed: stdin: {:?}\nstderr: {:?}",
+                command.stdout,
+                command.stderr
+            ))
+        }
     }
 }
 
 impl Inventory {
-    fn get_hosts(self) -> Result<Vec<Host>> {
-        todo!();
-        return Err(anyhow!("-1"))
+    pub fn get_hosts(self) -> Result<Vec<Host>> {
+        let mut hosts = Vec::new();
+        for item in self.data["all"]["children"].as_array().unwrap() {
+            let host_vars = &self.data["_meta"]["hostvars"][item.as_str().unwrap()];
+            let host = Host {
+                name: item.to_owned(),
+                data: host_vars.to_owned(),
+            };
+            hosts.push(host);
+        }
+        Ok(hosts)
     }
 }
 
 impl Host {
-    fn get_vars(self) -> Result<Value> {
-        todo!();
-        return Err(anyhow!("-1"))
+    pub fn get_vars(self) -> Result<Value> {
+        Ok(self.data)
     }
 }