+++ CHANGELOG.md
@@ -5,4 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025.05.07
+- Structs for Ansible `Inventory` and `Host` objects
+- `Load` trait for `Inventory` struct
+- `get_hosts()` method for `Inventory` struct
+- `get_vars()` method for `Host` struct
+- Basic test coverage
+
 ## [0.1.0] - 2025.05.07
+- Initial commit containing boilerplate Rust library code